fix(test): gate non-Linux cgroup test with //go:build !linux
On Linux, cgroup.NewManager succeeds, so the "not implemented" assertion
fails. Move the non-Linux NewManager test to cgroup_other_test.go with
//go:build !linux build tag.
